Move standard JobCatalog implementations to gobblin.runtime.job_catalog
