+++ gobblin-runtime/src/main/java/gobblin/runtime/api/GobblinInstanceLauncher.java
@@ -11,6 +11,8 @@
  */
 package gobblin.runtime.api;
 
+import com.google.common.util.concurrent.Service;
+
 import gobblin.annotation.Alpha;
 
 /**
@@ -18,7 +20,7 @@ import gobblin.annotation.Alpha;
  *  configurations, instantiates and runs the Gobblin instance driver.
  */
 @Alpha
-public interface GobblinInstanceLauncher extends StandardRunControls {
+public interface GobblinInstanceLauncher extends Service {
   /**
    * Creates a new Gobblin instance to run Gobblin jobs.
    * @throws IllegalStateException if {@link #isRunning()} is false.*/
