+++ gobblin-runtime/src/main/java/gobblin/runtime/job_catalog/CachingJobCatalog.java
@@ -9,7 +9,7 @@
  * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
  * CONDITIONS OF ANY KIND, either express or implied.
  */
-package gobblin.runtime.std;
+package gobblin.runtime.job_catalog;
 
 import java.net.URI;
 import java.util.Collection;
