+++ gobblin-runtime/src/main/java/gobblin/runtime/job_catalog/InMemoryJobCatalog.java
@@ -9,7 +9,7 @@
  * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
  * CONDITIONS OF ANY KIND, either express or implied.
  */
-package gobblin.runtime.std;
+package gobblin.runtime.job_catalog;
 
 import java.net.URI;
 import java.util.ArrayList;
@@ -27,7 +27,7 @@ import gobblin.runtime.api.JobCatalogListener;
 import gobblin.runtime.api.JobSpec;
 import gobblin.runtime.api.JobSpecNotFoundException;
 import gobblin.runtime.api.MutableJobCatalog;
-import gobblin.runtime.std.JobCatalogListenersList.AddJobCallback;
+import gobblin.runtime.job_catalog.JobCatalogListenersList.AddJobCallback;
 
 /**
  * Simple implementation of a Gobblin job catalog that stores all JobSpecs in memory. No persistence
